Add StyleCop to MySqlConnector project. Fixes #94
Disable some rules; fix the code for others.
