+++ src/MySqlConnector/project.json
@@ -4,11 +4,12 @@
 	"description": "A fully async MySQL ADO.NET provider for .NET and .NET Core",
 	"copyright": "Copyright 2016 Bradley Grainger",
 	"buildOptions": {
-		"warningsAsErrors": true
+		"warningsAsErrors": true,
+		"additionalArguments": [ "/ruleset:stylecop.ruleset" ]
 	},
 	"packOptions": {
 		"owners": [ "bgrainger" ],
-		"tags": [ "mysql", "async", "ado.net", "database", "netcore"  ],
+		"tags": [ "mysql", "async", "ado.net", "database", "netcore" ],
 		"releaseNotes": "Provide more information for unsupported auth plugins.",
 		"licenseUrl": "https://opensource.org/licenses/MIT",
 		"projectUrl": "https://github.com/bgrainger/MySqlConnector",
@@ -25,7 +26,11 @@
 				"System.Transactions": "4.0.0.0"
 			},
 			"dependencies": {
-				"System.Threading.Tasks.Extensions": "4.0.0"
+				"System.Threading.Tasks.Extensions": "4.0.0",
+				"StyleCop.Analyzers": {
+					"version": "1.0.0",
+					"type": "build"
+				}
 			}
 		},
 		"netstandard1.3": {
@@ -43,7 +48,11 @@
 				"System.Runtime.Extensions": "4.1.0",
 				"System.Security.Cryptography.Algorithms": "4.2.0",
 				"System.Threading": "4.0.11",
-				"System.Threading.Tasks.Extensions": "4.0.0"
+				"System.Threading.Tasks.Extensions": "4.0.0",
+				"StyleCop.Analyzers": {
+					"version": "1.0.0",
+					"type": "build"
+				}
 			}
 		}
 	}
